Update-WIP
I got most of the important resusable pieces in. All that's left is the actual implementation of STAC to GeoDID (pinning and anchoring). We could just use the STAC ID for the Method Specific Identifier for right now. I'm not implementing the Authentication yet and/or the GeoDoctype Handler yet. Just going to focus on scraping the STAC Item and converting it to a GeoDID with the assets and document stored on FFS