Update
We need to figure out a way to upload the GeoTiff to FFS. Ideas I had was to upload to IPFS via the IPFS client or upload to FFS or IPFS with the CLI. As of right now the pinning works for the smaller assets within the STAC item. The gRPC limit is 4194304 bytes, and the Cloud Optimized GeoTIFF is around 154251089. The following error was: (node:3653) UnhandledPromiseRejectionWarning: error code 8 - grpc: received message larger than max (154251089 vs. 4194304). It is also a current Powergate-client Issue @ https://github.com/textileio/js-powergate-client/issues/133 . They mention something about chunking the files, I'll look into it tomorrow.