Update
You can create a GeoDID now, but I just have to fix some things to make it so you can load in the future. The authentication still doesnt work yet either, but I think we can wait to incoporate it later when we actually have a server that can take care of decryption.